problemfixing

+++ docker-cm-dashing/rootfs/init/icinga_cert.sh
@@ -1,6 +1,6 @@
 
 ICINGA_API_PORT=${ICINGA_API_PORT:-5665}
-ICINGA_CERT_SERVICE=${ICINGA_CERT_SERVICE:-false}
+USE_CERT_SERVICE=${USE_CERT_SERVICE:-'false'}
 
 ICINGA_CERT_SERVICE_SERVER=
 ICINGA_CERT_SERVICE_PORT=
@@ -21,11 +21,12 @@ get_certificate() {
     return
   fi
 
-  if [ ${ICINGA_CERT_SERVICE} ]
+  if [ "${USE_CERT_SERVICE}" == "true" ]
   then
     echo ""
     echo " [i] we ask our cert-service for a certificate .."
 
+    set -x
     # generate a certificate request
     #
     code=$(curl \
@@ -33,10 +34,12 @@ get_certificate() {
       --silent \
       --request GET \
       --header "X-API-USER: ${ICINGA_CERT_SERVICE_API_USER}" \
-      --header "X-API-KEY: ${ICINGA_CERT_SERVICE_API_PASSWORD}" \
+      --header "X-API-PASSWORD: ${ICINGA_CERT_SERVICE_API_PASSWORD}" \
       --write-out "%{http_code}\n" \
       --output /tmp/request_${HOSTNAME}.json \
       http://${ICINGA_CERT_SERVICE_SERVER}:${ICINGA_CERT_SERVICE_PORT}${ICINGA_CERT_SERVICE_PATH}v2/request/${HOSTNAME})
+
+    set +x
 
     if ( [ $? -eq 0 ] && [ ${code} -eq 200 ] )
     then
@@ -58,7 +61,7 @@ get_certificate() {
         --silent \
         --request GET \
         --header "X-API-USER: ${ICINGA_CERT_SERVICE_API_USER}" \
-        --header "X-API-KEY: ${ICINGA_CERT_SERVICE_API_PASSWORD}" \
+        --header "X-API-PASSWORD: ${ICINGA_CERT_SERVICE_API_PASSWORD}" \
         --header "X-CHECKSUM: ${checksum}" \
         --write-out "%{http_code}\n" \
         --request GET \
@@ -101,12 +104,17 @@ get_certificate() {
       fi
     else
 
-      error=$(cat /tmp/request_${HOSTNAME}.json)
+      if [ -f /tmp/request_${HOSTNAME}.json ]
+      then
+        error=$(cat /tmp/request_${HOSTNAME}.json)
 
-      echo " [E] ${code} - the cert-service tell us a problem: '${error}'"
-      echo " [E] exit ..."
+        echo " [E] ${code} - the cert-service tell us a problem: '${error}'"
+        echo " [E] exit ..."
 
-      rm -f /tmp/request_${HOSTNAME}.json
+        rm -f /tmp/request_${HOSTNAME}.json
+      else
+        echo " [E] ${code} - the cert-service has an unknown problem."
+      fi
       exit 1
     fi
 
@@ -132,7 +140,7 @@ validate_local_ca() {
       --silent \
       --request GET \
       --header "X-API-USER: ${ICINGA_CERT_SERVICE_API_USER}" \
-      --header "X-API-KEY: ${ICINGA_CERT_SERVICE_API_PASSWORD}" \
+      --header "X-API-PASSWORD: ${ICINGA_CERT_SERVICE_API_PASSWORD}" \
       --write-out "%{http_code}\n" \
       --output /tmp/validate_ca_${HOSTNAME}.json \
       http://${ICINGA_CERT_SERVICE_SERVER}:${ICINGA_CERT_SERVICE_PORT}${ICINGA_CERT_SERVICE_PATH}v2/validate/${checksum})
@@ -161,6 +169,21 @@ extract_vars() {
 
   if [ ! -z "${ICINGA_CERT_SERVICE}" ]
   then
+
+#     echo "${ICINGA_CERT_SERVICE}" | json_verify -q 2> /dev/null
+#
+#     if [ $? -gt 0 ]
+#     then
+#       echo " [E] the ICINGA_CERT_SERVICE Environment ist not an json"
+#       exit 1
+#     fi
+#
+#     if ( [ "${ICINGA_CERT_SERVICE}" == "true" ] || [ "${ICINGA_CERT_SERVICE}" == "true" ] )
+#     then
+#       echo " [E] the ICINGA_CERT_SERVICE Environment must be an json, not true or false!"
+#       exit 1
+#     fi
+
     ICINGA_CERT_SERVICE_SERVER=$(echo "${ICINGA_CERT_SERVICE}" | jq --raw-output .server)
     ICINGA_CERT_SERVICE_PORT=$(echo "${ICINGA_CERT_SERVICE}" | jq --raw-output .port)
     ICINGA_CERT_SERVICE_PATH=$(echo "${ICINGA_CERT_SERVICE}" | jq --raw-output .path)
@@ -186,7 +209,7 @@ extract_vars() {
       [ ! -z ${ICINGA_CERT_SERVICE_API_PASSWORD} ]
     )
     then
-      ICINGA_CERT_SERVICE=true
+      USE_CERT_SERVICE="true"
     fi
   fi
 }
